wrote snapshot tests for app and videoPlayer components
