+++ client/src/app/app.test.js
@@ -1,0 +1,12 @@
+import React from 'react';
+import ReactDOM from 'react-dom';
+import { render, cleanup } from '@testing-library/react';
+import '@testing-library/jest-dom';
+import App from './app';
+
+afterEach(cleanup);
+
+it('matches snapshot', () => {
+  const { asFragment } = render(<App />);
+  expect(asFragment()).toMatchSnapshot();
+});
