+++ package.json
@@ -13,15 +13,14 @@
     "start": "node ./server/index.js",
     "start:dev": "nodemon index.js",
     "seed": "node database/seed.js",
-    "lint-react": "eslint ./client/src/**/*.js ./client/src/*.js"
+    "lint-react": "eslint ./client/src/**/*.js ./client/src/*.js",
+    "test": "jest"
   },
   "dependencies": {
-    "aws-sdk": "^2.666.0",
-    "body-parser": "^1.19.0",
+    "@testing-library/jest-dom": "^5.5.0",
+    "@testing-library/react": "^10.0.4",
     "express": "^4.17.1",
     "faker": "^4.1.0",
-    "fs": "0.0.1-security",
-    "image-downloader": "^3.5.0",
     "jquery": "^3.5.0",
     "jsdom": "^16.2.2",
     "lodash": "^4.17.15",
@@ -29,7 +28,8 @@
     "path": "^0.12.7",
     "prop-types": "^15.7.2",
     "react": "^16.13.1",
-    "react-dom": "^16.13.1"
+    "react-dom": "^16.13.1",
+    "react-testing-library": "^8.0.1"
   },
   "devDependencies": {
     "@babel/core": "^7.9.0",
@@ -44,7 +44,9 @@
     "eslint-plugin-react": "^7.19.0",
     "eslint-plugin-react-hooks": "^2.5.1",
     "jest": "^26.0.0",
+    "jest-dom": "^4.0.0",
     "nodemon": "^2.0.3",
+    "react-test-renderer": "^16.13.1",
     "webpack": "^4.43.0",
     "webpack-cli": "^3.3.11"
   }
