+++ public/index.html
@@ -7,8 +7,6 @@
   </head>
   <body>
     <div id="app">HTML!!!!!!!!!!</div>
-    <script src="https://unpkg.com/react@16/umd/react.development.js" crossorigin></script>
-    <script src="https://unpkg.com/react-dom@16/umd/react-dom.development.js" crossorigin></script>
     <script src="bundle.js"></script>
   </body>
 </html>
